Agregado manejadores de metaprogramacion
