+++ src/main/groovy/dsl/EmpresaDSL.groovy
@@ -1,13 +1,13 @@
 package dsl
 
-import groovy.lang.MetaClass;
-
 import org.joda.time.DateTime
-import org.joda.time.Hours
 
 import ar.edu.utn.tadp.empresa.Empresa.*
 import ar.edu.utn.tadp.propiedad.Propiedad
 import ar.edu.utn.tadp.requerimiento.Requerimiento
+import dsl.manejadores.ManejadorDeRecursos;
+import dsl.manejadores.ManejadorDeRequerimientos;
+import dsl.manejadores.ManejadorDeRoles
 
 class EmpresaDSL {
 	
@@ -18,9 +18,9 @@ class EmpresaDSL {
 	def reunion
 	def duracion
 	
-	def EmpresaDSL(unaEmpresa){
+	def EmpresaDSL(unaEmpresa, manejadores = [new ManejadorDeRoles(), new ManejadorDeRecursos(), new ManejadorDeRequerimientos()]){
 		empresa = unaEmpresa
-		initialize()
+		initialize(manejadores)
 	}
 	
 	def anfitrion(anfitrion){
@@ -64,32 +64,10 @@ class EmpresaDSL {
 		reunion.agregarTratamientoPorReplanificacion()
 	}
 	//++++++++++++++++++++++++++++++++++++++++++++++++++++++
-	//++ Personas ++++++++++++++++++++++++++++++++++++++++++
+	//++ Inicializacion de Manejadores ++++++++++++++++++++++++++++++++++++++++++
 	//++++++++++++++++++++++++++++++++++++++++++++++++++++++
-	def initialize() {
-		def otroRequerimiento = [ "sector" ]
-		
-		otroRequerimiento.each { value ->
-			this.getClass().metaClass."$value" << { String requerimiento -> agregarRequerimiento(cantidad, [new Propiedad(value,requerimiento)]); this} 
-		}
-		
-		def map = [
-			gerente:"Gerente",
-			programador:"programador",
-			liderTecnico:"Lider Tecnico",
-			projectLeader:"project leader",
-			diseniadorGrafico:"graphic designer"]
-		
-		map.each { key, value ->
-			this.getClass().metaClass."$key" << { -> agregarRequerimiento(cantidad, [new Propiedad("rol",value)]); this }
-			this.getClass().metaClass."$key" << { String proyecto -> agregarRequerimiento(cantidad, [new Propiedad("proyecto",proyecto), new Propiedad("rol",value)]); this } 
-		}
-		
-		def recursos = ["proyector", "notebook"]
-		
-		recursos.each { value -> 
-			this.getClass().metaClass."$value" << { -> agregarRequerimiento(cantidad, [new Propiedad("tipo",value)]); this }
-		} 
+	def initialize(manejadores) {
+		manejadores.each { manejador -> manejador.execute(this) }
 	}
 	
 	def agregarRequerimiento(cant, propiedades){
